+++ packages/auth/test/integration/flows/anonymous.test.ts
@@ -37,6 +37,7 @@ import {
   getTestInstance,
   randomEmail
 } from '../../helpers/integration/helpers';
+import { generateMiddlewareTests } from './middleware_test_generator';
 
 use(chaiAsPromised);
 
@@ -127,5 +128,9 @@ describe('Integration test: anonymous auth', () => {
         'auth/email-already-in-use'
       );
     });
+  });
+
+  generateMiddlewareTests(() => auth, () => {
+    return signInAnonymously(auth);
   });
 });