+++ packages/firestore/test/unit/local/query_engine.test.ts
@@ -19,12 +19,22 @@ import { expect } from 'chai';
 
 import { Timestamp } from '../../../src';
 import { User } from '../../../src/auth/user';
-import { LimitType, Query, queryWithLimit } from '../../../src/core/query';
+import {
+  LimitType,
+  Query,
+  queryWithAddedFilter,
+  queryWithAddedOrderBy,
+  queryWithLimit
+} from '../../../src/core/query';
 import { SnapshotVersion } from '../../../src/core/snapshot_version';
 import { View } from '../../../src/core/view';
 import { DocumentOverlayCache } from '../../../src/local/document_overlay_cache';
+import { IndexedDbPersistence } from '../../../src/local/indexeddb_persistence';
+import {
+  INDEXING_ENABLED,
+  INDEXING_SCHEMA_VERSION
+} from '../../../src/local/indexeddb_schema';
 import { LocalDocumentsView } from '../../../src/local/local_documents_view';
-import { MemoryIndexManager } from '../../../src/local/memory_index_manager';
 import { MutationQueue } from '../../../src/local/mutation_queue';
 import { Persistence } from '../../../src/local/persistence';
 import { PersistencePromise } from '../../../src/local/persistence_promise';
@@ -35,14 +45,17 @@ import { TargetCache } from '../../../src/local/target_cache';
 import {
   documentKeySet,
   DocumentMap,
-  newMutationMap
+  newMutationMap,
+  documentMap
 } from '../../../src/model/collections';
 import { Document, MutableDocument } from '../../../src/model/document';
 import { DocumentKey } from '../../../src/model/document_key';
 import { DocumentSet } from '../../../src/model/document_set';
 import {
+  IndexKind,
   IndexOffset,
-  indexOffsetComparator
+  indexOffsetComparator,
+  newIndexOffsetFromDocument
 } from '../../../src/model/field_index';
 import { Mutation } from '../../../src/model/mutation';
 import { debugAssert } from '../../../src/util/assert';
@@ -53,23 +66,19 @@ import {
   key,
   orderBy,
   query,
+  fieldIndex,
+  patchMutation,
+  setMutation,
   version
 } from '../../util/helpers';
 
-import { testMemoryEagerPersistence } from './persistence_test_helpers';
+import * as persistenceHelpers from './persistence_test_helpers';
+import { TestIndexManager } from './test_index_manager';
 
 const TEST_TARGET_ID = 1;
 
 const MATCHING_DOC_A = doc('coll/a', 1, { matches: true, order: 1 });
 const NON_MATCHING_DOC_A = doc('coll/a', 1, { matches: false, order: 1 });
-const PENDING_MATCHING_DOC_A = doc('coll/a', 1, {
-  matches: true,
-  order: 1
-}).setHasLocalMutations();
-const PENDING_NON_MATCHING_DOC_A = doc('coll/a', 1, {
-  matches: false,
-  order: 1
-}).setHasLocalMutations();
 const UPDATED_DOC_A = doc('coll/a', 11, { matches: true, order: 1 });
 const MATCHING_DOC_B = doc('coll/b', 1, { matches: true, order: 2 });
 const UPDATED_MATCHING_DOC_B = doc('coll/b', 11, { matches: true, order: 2 });
@@ -101,13 +110,48 @@ class TestLocalDocumentsView extends LocalDocumentsView {
   }
 }
 
-describe('QueryEngine', () => {
+describe('MemoryQueryEngine', async () => {
+  genericQueryEngineTest(
+    /* durable= */ false,
+    persistenceHelpers.testMemoryEagerPersistence
+  );
+});
+
+describe('IndexedDbQueryEngine', async () => {
+  if (!IndexedDbPersistence.isAvailable()) {
+    console.warn('No IndexedDB. Skipping IndexedDbQueryEngine tests.');
+    return;
+  }
+
+  let persistencePromise: Promise<Persistence>;
+  beforeEach(async () => {
+    persistencePromise = persistenceHelpers.testIndexedDbPersistence({
+      schemaVersion: INDEXING_SCHEMA_VERSION
+    });
+  });
+
+  genericQueryEngineTest(/* durable= */ true, () => persistencePromise);
+});
+
+/**
+ * Defines the set of tests to run against the memory and IndexedDB-backed
+ * query engine.
+ *
+ * @param durable Whether the provided persistence is backed by IndexedDB
+ * @param persistencePromise A factory function that returns an initialized
+ * persistence layer.
+ */
+function genericQueryEngineTest(
+  durable: boolean,
+  persistencePromise: () => Promise<Persistence>
+): void {
   let persistence!: Persistence;
   let remoteDocumentCache!: RemoteDocumentCache;
-  let mutationQueue!: MutationQueue;
   let documentOverlayCache!: DocumentOverlayCache;
   let targetCache!: TargetCache;
   let queryEngine!: QueryEngine;
+  let indexManager!: TestIndexManager;
+  let mutationQueue!: MutationQueue;
   let localDocuments!: TestLocalDocumentsView;
 
   /** Adds the provided documents to the query target mapping.  */
@@ -126,10 +170,11 @@ describe('QueryEngine', () => {
   function addDocument(...docs: MutableDocument[]): Promise<void> {
     return persistence.runTransaction('addDocument', 'readwrite', txn => {
       const changeBuffer = remoteDocumentCache.newChangeBuffer();
-      for (const doc of docs) {
-        changeBuffer.addEntry(doc);
-      }
-      return changeBuffer.apply(txn);
+      return PersistencePromise.forEach(docs, (doc: MutableDocument) =>
+        changeBuffer
+          .getEntry(txn, doc.key)
+          .next(() => changeBuffer.addEntry(doc))
+      ).next(() => changeBuffer.apply(txn));
     });
   }
 
@@ -206,16 +251,18 @@ describe('QueryEngine', () => {
   }
 
   beforeEach(async () => {
-    persistence = await testMemoryEagerPersistence();
+    persistence = await persistencePromise();
     targetCache = persistence.getTargetCache();
     queryEngine = new QueryEngine();
 
-    const indexManager = persistence.getIndexManager(User.UNAUTHENTICATED);
+    const underlyingIndexManager = persistence.getIndexManager(
+      User.UNAUTHENTICATED
+    );
     remoteDocumentCache = persistence.getRemoteDocumentCache();
-    remoteDocumentCache.setIndexManager(indexManager);
+    remoteDocumentCache.setIndexManager(underlyingIndexManager);
     mutationQueue = persistence.getMutationQueue(
       User.UNAUTHENTICATED,
-      indexManager
+      underlyingIndexManager
     );
     documentOverlayCache = persistence.getDocumentOverlayCache(
       User.UNAUTHENTICATED
@@ -224,9 +271,18 @@ describe('QueryEngine', () => {
       remoteDocumentCache,
       mutationQueue,
       documentOverlayCache,
-      new MemoryIndexManager()
+      underlyingIndexManager
     );
-    queryEngine.setLocalDocumentsView(localDocuments);
+    queryEngine.initialize(localDocuments, underlyingIndexManager);
+
+    indexManager = new TestIndexManager(persistence, underlyingIndexManager);
+  });
+
+  afterEach(async () => {
+    if (persistence.started) {
+      await persistence.shutdown();
+      await persistenceHelpers.clearTestPersistence();
+    }
   });
 
   it('uses target mapping for initial view', async () => {
@@ -248,8 +304,8 @@ describe('QueryEngine', () => {
     await addDocument(MATCHING_DOC_A, MATCHING_DOC_B);
     await persistQueryMapping(MATCHING_DOC_A.key, MATCHING_DOC_B.key);
 
-    // Add a mutated document that is not yet part of query's set of remote keys.
-    await addDocument(PENDING_NON_MATCHING_DOC_A);
+    // Add a mutation that is not yet part of query's set of remote keys.
+    await addMutation(patchMutation('coll/a', { 'matches': false }));
 
     const docs = await expectOptimizedCollectionQuery(() =>
       runQuery(query1, LAST_LIMBO_FREE_SNAPSHOT)
@@ -349,8 +405,9 @@ describe('QueryEngine', () => {
 
     // Add a query mapping for a document that matches, but that sorts below
     // another document due to a pending write.
-    await addDocument(PENDING_MATCHING_DOC_A);
-    await persistQueryMapping(PENDING_MATCHING_DOC_A.key);
+    await addDocument(MATCHING_DOC_A);
+    await addMutation(patchMutation('coll/a', { order: 1 }));
+    await persistQueryMapping(MATCHING_DOC_A.key);
 
     await addDocument(MATCHING_DOC_B);
 
@@ -368,8 +425,9 @@ describe('QueryEngine', () => {
     );
     // Add a query mapping for a document that matches, but that sorts below
     // another document due to a pending write.
-    await addDocument(PENDING_MATCHING_DOC_A);
-    await persistQueryMapping(PENDING_MATCHING_DOC_A.key);
+    await addDocument(MATCHING_DOC_A);
+    await addMutation(patchMutation('coll/a', { order: 2 }));
+    await persistQueryMapping(MATCHING_DOC_A.key);
 
     await addDocument(MATCHING_DOC_B);
 
@@ -432,7 +490,7 @@ describe('QueryEngine', () => {
     await persistQueryMapping(key('coll/a'), key('coll/b'));
 
     // Update "coll/a" but make sure it still sorts before "coll/b"
-    await addDocument(doc('coll/a', 1, { order: 2 }).setHasLocalMutations());
+    await addMutation(patchMutation('coll/a', { order: 2 }));
 
     // Since the last document in the limit didn't change (and hence we know
     // that all documents written prior to query execution still sort after
@@ -458,7 +516,117 @@ describe('QueryEngine', () => {
     );
     verifyResult(docs, [MATCHING_DOC_A]);
   });
-});
+
+  if (!durable) {
+    return;
+  }
+
+  if (!INDEXING_ENABLED) {
+    return;
+  }
+
+  it('combines indexed with non-indexed results', async () => {
+    debugAssert(durable, 'Test requires durable persistence');
+
+    const doc1 = doc('coll/a', 1, { 'foo': true });
+    const doc2 = doc('coll/b', 2, { 'foo': true });
+    const doc3 = doc('coll/c', 3, { 'foo': true });
+    const doc4 = doc('coll/d', 3, { 'foo': true }).setHasLocalMutations();
+
+    await indexManager.addFieldIndex(
+      fieldIndex('coll', { fields: [['foo', IndexKind.ASCENDING]] })
+    );
+
+    await addDocument(doc1);
+    await addDocument(doc2);
+    await indexManager.updateIndexEntries(documentMap(doc1, doc2));
+    await indexManager.updateCollectionGroup(
+      'coll',
+      newIndexOffsetFromDocument(doc2)
+    );
+
+    await addDocument(doc3);
+    await addMutation(setMutation('coll/d', { 'foo': true }));
+
+    const queryWithFilter = queryWithAddedFilter(
+      query('coll'),
+      filter('foo', '==', true)
+    );
+    const results = await expectOptimizedCollectionQuery(() =>
+      runQuery(queryWithFilter, SnapshotVersion.min())
+    );
+
+    verifyResult(results, [doc1, doc2, doc3, doc4]);
+  });
+
+  it('uses partial index for limit queries', async () => {
+    debugAssert(durable, 'Test requires durable persistence');
+
+    const doc1 = doc('coll/1', 1, { 'a': 1, 'b': 0 });
+    const doc2 = doc('coll/2', 1, { 'a': 1, 'b': 1 });
+    const doc3 = doc('coll/3', 1, { 'a': 1, 'b': 2 });
+    const doc4 = doc('coll/4', 1, { 'a': 1, 'b': 3 });
+    const doc5 = doc('coll/5', 1, { 'a': 2, 'b': 3 });
+    await addDocument(doc1, doc2, doc3, doc4, doc5);
+
+    await indexManager.addFieldIndex(
+      fieldIndex('coll', { fields: [['a', IndexKind.ASCENDING]] })
+    );
+    await indexManager.updateIndexEntries(
+      documentMap(doc1, doc2, doc3, doc4, doc5)
+    );
+    await indexManager.updateCollectionGroup(
+      'coll',
+      newIndexOffsetFromDocument(doc5)
+    );
+
+    const q = queryWithLimit(
+      queryWithAddedFilter(
+        queryWithAddedFilter(query('coll'), filter('a', '==', 1)),
+        filter('b', '==', 1)
+      ),
+      3,
+      LimitType.First
+    );
+    const results = await expectOptimizedCollectionQuery(() =>
+      runQuery(q, SnapshotVersion.min())
+    );
+
+    verifyResult(results, [doc2]);
+  });
+
+  it('re-fills indexed limit queries', async () => {
+    debugAssert(durable, 'Test requires durable persistence');
+
+    const doc1 = doc('coll/1', 1, { 'a': 1 });
+    const doc2 = doc('coll/2', 1, { 'a': 2 });
+    const doc3 = doc('coll/3', 1, { 'a': 3 });
+    const doc4 = doc('coll/4', 1, { 'a': 4 });
+    await addDocument(doc1, doc2, doc3, doc4);
+
+    await indexManager.addFieldIndex(
+      fieldIndex('coll', { fields: [['a', IndexKind.ASCENDING]] })
+    );
+    await indexManager.updateIndexEntries(documentMap(doc1, doc2, doc3, doc4));
+    await indexManager.updateCollectionGroup(
+      'coll',
+      newIndexOffsetFromDocument(doc4)
+    );
+
+    await addMutation(patchMutation('coll/3', { 'a': 5 }));
+
+    const q = queryWithLimit(
+      queryWithAddedOrderBy(query('coll'), orderBy('a')),
+      3,
+      LimitType.First
+    );
+    const results = await expectOptimizedCollectionQuery(() =>
+      runQuery(q, SnapshotVersion.min())
+    );
+
+    verifyResult(results, [doc1, doc2, doc4]);
+  });
+}
 
 function verifyResult(actualDocs: DocumentSet, expectedDocs: Document[]): void {
   for (const doc of expectedDocs) {