+++ packages/app-check-compat/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @firebase/app-check-compat
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [[`9c5c9c36d`](https://github.com/firebase/firebase-js-sdk/commit/9c5c9c36da80b98b73cfd60ef2e2965087e9f801)]:
+  - @firebase/util@1.6.0
+  - @firebase/app-check@0.5.8
+  - @firebase/component@0.5.14
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [[`38da5d9be`](https://github.com/firebase/firebase-js-sdk/commit/38da5d9be4a001038548947a85c9b612e5275e32)]:
+  - @firebase/app-check@0.5.7
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [[`e9e5f6b3c`](https://github.com/firebase/firebase-js-sdk/commit/e9e5f6b3ca9d61323b22f87986d9959f5297ec59)]:
+  - @firebase/util@1.5.2
+  - @firebase/app-check@0.5.6
+  - @firebase/component@0.5.13
+
 ## 0.2.5
 
 ### Patch Changes